auto-improve: [#799 Step 3/6] Move `parse.py` and `publish.py` into `cai_lib/`

Refs damien-robotsix/robotsix-cai#820
